requirements.txt
flask==2.3.3
flask_cors==4.0.0
ipinfo==4.1.0
gunicorn==21.2.0
google-cloud-firestore==2.16.0
user-agents==2.2.0